moved .cpp and .hpp of analog and lookup
