--- Src/ST-LIB_LOW/Sensors/LinearSensor/AnalogSensor.cpp
@@ -1,3 +1,0 @@
-
-#include "ADC/ADC.hpp"
-#include "ST-LIB_LOW/Sensors/LinearSensor/AnalogSensor.hpp"
